Add files via upload
This was a group project done for the course - Data Analytics for Engg. Sys.  -  but all the code files (R Studio) scripts and TABLEAU Files are owned and developed by me. The logistics regression and linear regression scripts were developed as a team with my team mates in the project. Refer to the PPT file for better understanding on the results and conclusions drawn from the analyses conducted.